bring back new conversation list item on desktop
