+++ shared/chat/inbox/row/index.js
@@ -10,6 +10,7 @@ import {Box} from '../../../common-adapters'
 import {globalStyles, globalMargins} from '../../../styles'
 import {isMobile} from '../../../constants/platform'
 import * as Constants from '../../../constants/chat'
+import NewConversation from './new-conversation'
 
 type MakeRowOptions = {
   channelname: string,
@@ -71,15 +72,16 @@ const makeRow = (options: MakeRowOptions) => {
           />
         )
       case 'small':
-        return (
-          <SmallTeam
-            key={options.conversationIDKey}
-            conversationIDKey={options.conversationIDKey}
-            channelname={options.channelname}
-            isActiveRoute={options.isActiveRoute}
-            teamname={options.teamname}
-          />
-        )
+        // $FlowIssue conversationIDKey is totally a string
+        return Constants.isPendingConversationIDKey(options.conversationIDKey) && !isMobile
+          ? <NewConversation />
+          : <SmallTeam
+              key={options.conversationIDKey}
+              conversationIDKey={options.conversationIDKey}
+              channelname={options.channelname}
+              isActiveRoute={options.isActiveRoute}
+              teamname={options.teamname}
+            />
     }
   }
   return new Error(`Unhandled row type ${options.type}`)
