+++ shared/chat/inbox/row/new-conversation.js
@@ -1,0 +1,53 @@
+// @flow
+import React, {PureComponent} from 'react'
+import {Text, Icon} from '../../../common-adapters'
+import {globalStyles, globalColors, globalMargins} from '../../../styles'
+
+class NewConversation extends PureComponent<{}> {
+  render() {
+    return (
+      <div
+        style={{
+          ...globalStyles.flexBoxRow,
+          alignItems: 'center',
+          backgroundColor: globalColors.blue,
+          flexShrink: 0,
+          minHeight: 48,
+        }}
+      >
+        <div
+          style={{
+            ...globalStyles.flexBoxRow,
+            ...globalStyles.clickable,
+            alignItems: 'center',
+          }}
+        >
+          <div
+            style={{
+              backgroundColor: 'rgba(255, 255, 255, 0.2)',
+              borderRadius: globalMargins.large,
+              height: globalMargins.large,
+              marginLeft: globalMargins.tiny,
+              marginRight: globalMargins.tiny,
+              padding: globalMargins.tiny,
+              width: globalMargins.large,
+            }}
+          >
+            <Icon
+              type="iconfont-chat"
+              style={{
+                color: globalColors.blue,
+                fontSize: 24,
+                marginLeft: 1,
+                marginTop: 1,
+              }}
+            />
+          </div>
+          <Text style={{color: globalColors.white}} type="BodySemibold">New conversation</Text>
+        </div>
+      </div>
+    )
+  }
+}
+
+export default NewConversation
